backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# RSS & Web Scraping
feedparser==6.0.10
newspaper3k==0.2.8
lxml==4.9.3
lxml-html-clean==0.4.3
beautifulsoup4==4.12.2
requests==2.31.0

# NLP & AI/ML
transformers==4.35.2
torch==2.1.1
sentencepiece==0.2.1
tokenizers==0.15.0
huggingface-hub==0.19.4
safetensors==0.4.1

# Translation
deep-translator==1.11.4
langdetect==1.0.9

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
PyYAML==6.0.1
python-dateutil==2.8.2
pytz==2023.3

# Email
aiosmtplib==3.0.1

# OpenAI & RAG (Optional)
openai==1.3.7
tiktoken==0.12.0

# Vector Store
faiss-cpu==1.7.4
numpy==1.24.3

# Testing & Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
